parameter implemented

+++ admin_view.php
@@ -28,7 +28,7 @@
 
                 <div class="row-fluid">
 
-                  <a href="new_casting.php" class="btn btn-info">
+                  <a href="new_casting.php" class="btn btn-primary">
 
                     <i class="icon-plus icon-white"></i> Add Casting
 
@@ -60,22 +60,18 @@
 
                     $result = mysql_query($query);
 
-                    while($row = mysql_fetch_array($result)) {
-
-                      echo "<tr>";
-
-                      echo "<td>".$row['title']."</td>";
-
-                      echo "<td>";
-
-                      echo "<a href=\"edit_casting.php?id=".$row['id']."\" class=\"btn\">Edit</a> &nbsp";
-
-                      echo "<a href=\"delete_casting.php?id=".$row['id']."\" class=\"btn btn-danger\">Delete</a>";
-
-                      echo "</td>";
-
-                      echo "</tr>";
-
+                    if(mysql_num_rows($result)>0){
+                      while($row = mysql_fetch_array($result)) {
+                        echo "<tr>";
+                        echo "<td>".$row['title']."</td>";
+                        echo "<td>";
+                        echo "<a href=\"edit_casting.php?id=".$row['id']."\" class=\"btn btn-info\">Edit</a> &nbsp";
+                        echo "<a href=\"delete_casting.php?id=".$row['id']."\" class=\"btn btn-danger\">Delete</a>";
+                        echo "</td>";
+                        echo "</tr>";
+                      }
+                    } else {
+                      echo "<tr><td colspan=\"2\"> No record to display </td></tr>";
                     }
 
                   ?>
